requirements.txt
gunicorn
flask
python-dotenv
flask-mail
